Moved muxing code outside BlockSealingBidderFactory
