+++ config-live-example.toml
@@ -27,8 +27,6 @@ dry_run_validation_url = "http://localhost:8545"
 
 ignore_cancellable_orders = true
 
-max_concurrent_seals = 4
-
 # genesis_fork_version = "0x00112233"
 
 sbundle_mergeable_signers = []
