+++ crates/rbuilder/src/live_builder/block_output/block_sealing_bidder_factory.rs
@@ -6,12 +6,8 @@ use crate::{
     live_builder::payload_events::MevBoostSlotData,
     provider::StateProviderFactory,
 };
-use ahash::HashMap;
-use alloy_primitives::{TxHash, U256};
-use std::{
-    fmt::Debug,
-    sync::{Arc, Mutex},
-};
+use alloy_primitives::U256;
+use std::{fmt::Debug, sync::Arc};
 use tracing::error;
 
 use super::{
@@ -131,54 +127,6 @@ where
     }
 }
 
-struct BestBlockState {
-    /// For each algo name we store the last block so on each update we can look for the best block among all algos.
-    /// We should replace String for "real" type identifying the algo.
-    algorithm_to_last_block: HashMap<String, BiddableUnfinishedBlock>,
-    /// Works as an id for the last block so we don't send the same twice since the same block could be generated by different algos.
-    last_sent_block_txs: Option<Vec<TxHash>>,
-}
-
-impl BestBlockState {
-    fn new() -> Self {
-        Self {
-            algorithm_to_last_block: Default::default(),
-            last_sent_block_txs: None,
-        }
-    }
-
-    /// Given a new BiddableUnfinishedBlock updates the best per algo and return the best one if
-    /// changes from last time.
-    fn update_best_block(
-        &mut self,
-        biddable_block: BiddableUnfinishedBlock,
-    ) -> Option<BiddableUnfinishedBlock> {
-        self.algorithm_to_last_block.insert(
-            biddable_block.block().builder_name().to_string(),
-            biddable_block,
-        );
-        let (_, best_biddable_block) = self
-            .algorithm_to_last_block
-            .iter()
-            .max_by_key(|(_, bb)| bb.true_block_value())
-            .unwrap();
-        let best_block_txs = best_biddable_block
-            .block()
-            .built_block_trace()
-            .transaction_hashes();
-        if self
-            .last_sent_block_txs
-            .as_ref()
-            .is_none_or(|txs| *txs != best_block_txs)
-        {
-            self.last_sent_block_txs = Some(best_block_txs);
-            Some(best_biddable_block.clone())
-        } else {
-            None
-        }
-    }
-}
-
 /// Helper object containing the bidder.
 /// It just forwards new blocks and new competitions bids (via SlotBidderToBidValueObs) to the bidder.
 struct BlockSealingBidder {
@@ -189,8 +137,6 @@ struct BlockSealingBidder {
     /// We forward best block and competition bids to the bidder.
     /// It will bid with the BidMaker it received on creation.
     bidder: Arc<dyn SlotBidder>,
-    /// needed state to check and update the best block.
-    best_block_state: Mutex<BestBlockState>,
 }
 
 impl BlockSealingBidder {
@@ -214,21 +160,13 @@ impl BlockSealingBidder {
             bid_value_source_to_unsubscribe: slot_bidder_to_bid_value_obs,
             competition_bid_value_source,
             bidder,
-            best_block_state: Mutex::new(BestBlockState::new()),
         }
     }
 }
 
 impl UnfinishedBlockBuildingSink for BlockSealingBidder {
     fn new_block(&self, biddable_block: BiddableUnfinishedBlock) {
-        if let Some(best_biddable_block) = self
-            .best_block_state
-            .lock()
-            .unwrap()
-            .update_best_block(biddable_block)
-        {
-            self.bidder.new_block(best_biddable_block);
-        }
+        self.bidder.new_block(biddable_block);
     }
     fn can_use_suggested_fee_recipient_as_coinbase(&self) -> bool {
         self.bidder.can_use_suggested_fee_recipient_as_coinbase()
